+++ packages/components/src/components/autocomplete/autocomplete.browser.e2e.tsx
@@ -15,6 +15,7 @@ import {
   t9n,
   disabled,
   formAssociated,
+  globalProps,
   openClose,
   topLayer,
   accessible,
@@ -28,6 +29,18 @@ import type { Autocomplete } from "./autocomplete";
 import { CSS, SLOTS } from "./resources";
 
 mockConsole();
+
+describe("global props", () => {
+  globalProps(
+    () => mount<Autocomplete>(<calcite-autocomplete label="Items" />),
+    () => page.getBySelector("calcite-input"),
+    {
+      autofocus: true,
+      enterKeyHint: "search",
+      inputMode: "search",
+    },
+  );
+});
 
 describe("accessible", () => {
   describe("default", () => {