+++ packages/components/src/components/block/block.browser.e2e.tsx
@@ -30,7 +30,7 @@ mockConsole();
 describe("accessible", () => {
   accessible(() =>
     mount(
-      <calcite-block collapsible description="description" expanded heading="heading">
+      <calcite-block description="description" expandable expanded heading="heading">
         <div>content</div>
       </calcite-block>,
     ),
@@ -43,6 +43,10 @@ describe("defaults", () => {
     [
       {
         propertyName: "collapsible",
+        defaultValue: false,
+      },
+      {
+        propertyName: "expandable",
         defaultValue: false,
       },
       {
@@ -94,7 +98,7 @@ describe("setFocus", () => {
     focusable(
       () =>
         mount(
-          <calcite-block collapsible description="summary" expanded heading="Heading">
+          <calcite-block description="summary" expandable expanded heading="Heading">
             <calcite-block-section expanded text="input block-section">
               <calcite-input
                 icon="form-field"
@@ -140,6 +144,10 @@ describe("reflects", () => {
     [
       {
         propertyName: "collapsible",
+        value: true,
+      },
+      {
+        propertyName: "expandable",
         value: true,
       },
       {
@@ -242,7 +250,7 @@ describe("top layer placement", () => {
 });
 
 describe("disabled", () => {
-  disabled(() => mount(<calcite-block collapsible description="description" heading="heading" />));
+  disabled(() => mount(<calcite-block description="description" expandable heading="heading" />));
 });
 
 describe("theme", () => {
@@ -251,8 +259,8 @@ describe("theme", () => {
       () =>
         mount(
           <calcite-block
-            collapsible
             description="description"
+            expandable
             expanded
             heading="heading"
             icon-end="pen"
@@ -264,6 +272,9 @@ describe("theme", () => {
           </calcite-block>,
         ),
       {
+        "--calcite-block-background-color": {
+          targetProp: "backgroundColor",
+        },
         "--calcite-block-border-color": {
           targetProp: "borderColor",
         },
@@ -277,10 +288,6 @@ describe("theme", () => {
             targetProp: "paddingInline",
           },
         ],
-        "--calcite-block-header-background-color": {
-          shadowSelector: `.${CSS.toggle}`,
-          targetProp: "backgroundColor",
-        },
         "--calcite-block-header-background-color-hover": {
           shadowSelector: `.${CSS.toggle}`,
           targetProp: "backgroundColor",
@@ -335,8 +342,8 @@ describe("theme", () => {
       () =>
         mount(
           <calcite-block
-            collapsible
             description="description"
+            expandable
             expanded
             heading="heading"
             icon-end="pen"
@@ -348,6 +355,10 @@ describe("theme", () => {
           </calcite-block>,
         ),
       {
+        "--calcite-block-header-background-color": {
+          shadowSelector: `.${CSS.toggle}`,
+          targetProp: "backgroundColor",
+        },
         "--calcite-block-padding": [
           {
             shadowSelector: `section.${CSS.content}`,
@@ -393,7 +404,7 @@ describe("theme", () => {
   describe("toggleDisplay", () => {
     it("should toggle the expanded state when the toggleDisplay is switch", async () => {
       const { el } = await mount(
-        <calcite-block collapsible heading="heading" toggle-display="switch">
+        <calcite-block expandable heading="heading" toggle-display="switch">
           <div>content</div>
         </calcite-block>,
       );