+++ packages/components/src/components/input-number/input-number.browser.e2e.tsx
@@ -9,6 +9,7 @@ import {
   disabled,
   focusable,
   formAssociated,
+  globalProps,
   hidden,
   internalLabel,
   reflects,
@@ -23,7 +24,22 @@ import { CSS as ClearButtonCSS } from "../functional/ClearButton";
 import { CSS as InlineEditableControlsCSS } from "../functional/InlineEditableControls";
 import { defaultValidity } from "../../tests/commonTests/browser/defaults";
 import { CSS, DIRECTION, NUDGE_DELAY_IN_MS } from "./resources";
-import { InputNumber } from "./input-number";
+import type { InputNumber } from "./input-number";
+
+describe("global props", () => {
+  globalProps(
+    () => mount("calcite-input-number"),
+    () => page.getByRole("textbox"),
+    {
+      autofocus: true,
+      enterKeyHint: "done",
+      inputMode: "numeric",
+    },
+    {
+      inputMode: "decimal",
+    },
+  );
+});
 
 describe("defaults", () => {
   defaults(