+++ packages/components/src/components/input-time-picker/input-time-picker.tsx
@@ -435,7 +435,7 @@ export class InputTimePicker extends LitElement implements LabelableComponent, T
     }
   }
 
-  private mouseDownHandler(event): void {
+  private mouseDownHandler(event: MouseEvent): void {
     if (this.showPlaceholder) {
       event.preventDefault();
       this.setFocus();